+++ CHANGELOG.md
@@ -20,6 +20,7 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 
 * mTLS support.
 * `expose.mtls_ca_cert` setting.
+* `stats` command.
 * `destroy` command.
 * Show gufo-thor version on startup.